Started building the framework
The Frame class will handle what is displayed on the frame. Later, it
will be used to swap "states" (from settings to level editor or to the
game etc). Right now it adds the Game class (which is a canvas).

The Game class is the class which handles the game once it is started.

The Screen class will later work as a class where everything that should
be rendered is put together into a pixel array. Right not it just
renders tiles with a size of 16 pixels. The colors are random.

The InputHandler class handles basic input. More keys will be added in
the future. With the WASD keys or the arrow keys the tiles on the screen
are moveable.