Added sprites/sprite sheets, tiles and started on levels
The sprites are loaded from "spitesheet.png" in the "res"-folder. Each
pixel data is then stored into a pixel array. The Sprite class grabs
individuall sprites from the SpriteSheet class.

The subclasses of the Tile class will later get properties (such as if
it is solid, luminous, breakable). Right now it is assigned a sprite and
renders it to the Screen class' pixel array. Later, the Tile class will
only be a template for other tiles.

The Level class calls every individual tile's rendered method. The class
gets the screens corner pins and renders only the tiles that are
visible. Right now there is an intentional bug which makes it easier to
understand how the tiles are rendered.

The tiles are generated randomly between grass tiles (25 %) and empty
tiles (75 %). The level is currently 64 x 64 tiles large. When the
player reaches the end of the level, empty tiles are being rendered.