Added a new AI
The Stalker AI can currently see the player at an infinite distance when
there is no solid tile blocking its sight. When it sees the player, it
starts moving towards him/her. Once the Stalker AI has seen the player,
it continues to follow him/her (even when the player is not visible by
the AI) until it collides.