Added collision detection for AIs
The projectiles now collide with the AIs. They can also collide with
eachother. When the projectile hits the AI it removes a certain damage
from the AI's health. When this drops below 0 the AI dies. At the
moment, the AIs have two health points. Each projectile damages the AI
with one health point.

Later, collision detected with the player and the AIs need to be added.