Added health bars and non-square sprites
The health bars are rendered over everything else. Currently they are
shown for six seconds after its mob has been damaged. The colors of the
health bars are calculated by comparing the max health of a mob with its
current health.

The yellow bar is a feature which may later be used so that the mobs can
regain some of their lost health.

Non-square sprites have also been added.

I accidentally rolled back this commit so I rewrote the code. Hopefully
there are no bugs in it, but something may have been missed.