Added random levels
The idea behind the random level generation is to create rooms connected
with paths. This is accomplished by randomly placing rectangles on a set
range. By these rectangles some are randomly removed. The others get
connected with paths. Based on the rooms/rectangles, an array of tiles
is generated. Where a room or path was, a grass or flower tile is placed
(grass: 8/9, flower: 1/9).

Mob spawning has yet to be implemented on the random levels.

The mobs are now randomly spawned on non-solid tiles. Also, the Level
class now clears its list of mobs by itself. This prevents concurrent
modification exceptions to occur and removes the bug where some mobs did
not tick due to lists being changed.